跳跃表：修改概率计算的浮点数精度，float64改为float32

+++ tools/datastructure/skiplist/common_test.go
@@ -22,7 +22,7 @@ import "testing"
 func Test_randomLevel(t *testing.T) {
 	type args struct {
 		maxLevel           int
-		levelUpProbability float64
+		levelUpProbability float32
 	}
 	tests := []struct {
 		name string
